Remove unused view attachment details components
